add token-counter and tune chat message list
